+++ src/Exception/Component/UnknownComponentKeyException.php
@@ -6,4 +6,16 @@ use Torr\Storyblok\Exception\StoryblokException;
 
 final class UnknownComponentKeyException extends \InvalidArgumentException implements StoryblokException
 {
+	/**
+	 * @inheritDoc
+	 */
+	public function __construct (
+		string $message,
+		public readonly string $componentKey,
+		?\Throwable $previous = null,
+	)
+	{
+		parent::__construct($message, 0, $previous);
+	}
+
 }